WIP make save effect
So I added savesData predicate to Input and make Result check if the
action saves data and if so create an effect that sends itself a Save
message and does nothing.

This might be an odd choice. Perhaps the effect shoulds save the data
and then send something to the address of the list of results saying ...
But i;m getting ahead of myself. Because thats only for the case
where were updating the placeholder result.

Anyway this doesnt compile yet because Record.update needs an Address
Record.action to instantiate it. And we don't yet have the address
beause we use StartApp to create the mailbox so the next thing is to
create it explicityly.