Save Result only if field value changed
Pleasingly this change suggested that the conditino for saving results
is not to do with the kind of message being sent to the Input, but
whether the .vlaue changed.  Latch (which we were testing for before)is
the message that updates the value. Now I was going to check if the
value changed on top of using latch -- to avoid the case where you press
endter but no change to the value.  But this is redundant! If the value
changed, you must have used Latch. And if we ever aadded another way to
update the value, then we would want to know about that too.

At which point we  no longer need to be able to see into the
constructors for Input.Msg to decide whether to save, we just need
to see the .value of the Input.Model. Which we can because
its just a type alias for a record. So we can safely move the save
condition test out of Input and into the client.

Delightful.